Add ability to specify client id
Premium plan users use a Client ID instead of the API Key. This
adds support for a boolean flag that states whether the specified
key is a regular key or a client id. The flag should default to
false and is not required, thus existing implementations should
not break and this should be backward compatible.

See here for details:
https://developers.google.com/maps/documentation/javascript/get-api-key